+++ src/lib/helpers/typeahead.tsx
@@ -19,15 +19,27 @@ const convertAutoCompleteOptionsToStringArray = (options: TypeaheadOption[] | un
   return (options as LabelValueOption[]).map((option) => option.value) as string[];
 };
 
-const getSingleAutoCompleteValue = (options: TypeaheadOption[], value: LabelValueOption["value"]): TypeaheadOption[] =>
-  value ? options.filter((x) => (typeof x === "string" ? x == value : x.value == value)) : [];
+const getSingleAutoCompleteValue = (options: TypeaheadOption[], value: string | number | undefined): TypeaheadOption[] => {
+  if (value == undefined) {
+    return [];
+  }
+  return options.filter((x) =>
+    // loose equality check to handle different types between form value and option value
+    typeof x === "string" ? x == value : x.value == value,
+  );
+};
 
-const getMultipleAutoCompleteValue = (options: TypeaheadOption[], value: (string | number)[] | undefined): TypeaheadOption[] =>
-  value
-    ? options.filter((x) =>
-        typeof x === "string" || typeof x === "number" ? value.includes(x) : value.includes(x.value as string | number),
-      )
-    : [];
+const getMultipleAutoCompleteValue = (options: TypeaheadOption[], value: (string | number)[] | undefined): TypeaheadOption[] => {
+  if (value == undefined) {
+    return [];
+  }
+  return options.filter((x) =>
+    typeof x === "string"
+      ? value.includes(x)
+      : // ensure that form values matches options values even if they are of different types
+        value.map(String).includes(String(x.value as string | number)),
+  );
+};
 
 const getUniqueOptions = (source: TypeaheadOption[], comparison: TypeaheadOption[]): TypeaheadOption[] => {
   const comparisonSet = new Set(comparison.map((x) => (typeof x === "string" ? x : x.value)));